Include whathever GL utilities library is built
This depends on whether Qt builds with desktop GL or GLES2. We used to
build it with GLES2 on ARM otherwise desktop GL but that doesn't work
with qtwayland-egl so now we always build with GLES2 support.

We just shouldn't change kwin files list every time we deal with this
or similar issues.

A glob resolve the problem.